+++ ChangeLog.md
@@ -18,6 +18,8 @@
 - Added the `CopyArticleToClipboard` command for copying the current
   article's URL to the clipboard.
   ([#43](https://github.com/davep/oldnews/pull/43))
+- Added the `Copy` command for copying a URL to the clipboard depending on
+  the current context. ([#43](https://github.com/davep/oldnews/pull/43))
 
 ## v0.3.0